Moved gui into one class and started wiring up the databinding
